refactor RecipePage
