+++ packages/frontend/src/components/recipes/index.ts
@@ -1,1 +1,2 @@
-export { default as RecipeAddDialog } from 'src/components/recipes/RecipesAddDialog.vue';
+export { default as RecipesAddDialog } from 'src/components/recipes/RecipesAddDialog.vue';
+export { default as RecipeCard } from 'src/components/recipes/RecipeCard.vue';
